Refactor TopLevelCommandManager to use IExtensionService implementations
Extract extension loading from TopLevelCommandManager into separate
IExtensionService implementations:

- BuiltInExtensionService: wraps in-proc ICommandProvider instances from DI
- WinRTExtensionService: manages out-of-process WinRT AppExtension providers

TLCM now receives IEnumerable<IExtensionService> and orchestrates them
uniformly. This reduces TLCM from ~900 to ~700 lines and separates
provider-specific concerns (timeout, retry, catalog events) from
command orchestration.

Additional fixes:
- Break circular DI dependency: BuiltInsCommandProvider no longer needs
  IRootPageService. Uses GoHomeDockCommand (returns CommandResult.GoHome())
  with onBeforeShowConfirmation callback to show palette at dock position.
- Restore two-phase loading: PreLoadAsync loads only built-ins (instant),
  PostLoadRootPageAsync loads WinRT extensions on background thread.
- Add try/finally around IsLoading to prevent stuck loading indicator.
