+++ src/modules/cmdpal/Microsoft.CmdPal.UI.ViewModels/Commands/BuiltInsCommandProvider.cs
@@ -2,6 +2,7 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.CmdPal.UI.ViewModels.Commands;
 using Microsoft.CommandPalette.Extensions;
 using Microsoft.CommandPalette.Extensions.Toolkit;
 
@@ -18,8 +19,7 @@ public sealed partial class BuiltInsCommandProvider : CommandProvider
     private readonly FallbackReloadItem _fallbackReloadItem = new();
     private readonly FallbackLogItem _fallbackLogItem = new();
     private readonly NewExtensionPage _newExtension = new();
-
-    private readonly IRootPageService _rootPageService;
+    private readonly GoHomeDockCommand _goHomeDockCommand = new();
 
     public override ICommandItem[] TopLevelCommands() =>
         [
@@ -41,22 +41,16 @@ public sealed partial class BuiltInsCommandProvider : CommandProvider
             _fallbackLogItem,
         ];
 
-    public BuiltInsCommandProvider(IRootPageService rootPageService)
+    public BuiltInsCommandProvider()
     {
         Id = "com.microsoft.cmdpal.builtin.core";
         DisplayName = Properties.Resources.builtin_display_name;
         Icon = IconHelpers.FromRelativePath("Assets\\Square44x44Logo.altform-unplated_targetsize-256.png");
-
-        _rootPageService = rootPageService;
     }
 
     public override ICommandItem[]? GetDockBands()
     {
-        var rootPage = _rootPageService.GetRootPage();
-        List<ICommandItem> bandItems = new();
-        bandItems.Add(new WrappedDockItem(rootPage, Properties.Resources.builtin_command_palette_title));
-
-        return bandItems.ToArray();
+        return [new WrappedDockItem(_goHomeDockCommand, Properties.Resources.builtin_command_palette_title)];
     }
 
     public override void InitializeWithHost(IExtensionHost host) => BuiltinsExtensionHost.Instance.Initialize(host);
