+++ src/modules/cmdpal/Microsoft.CmdPal.UI.ViewModels/ShellViewModel.cs
@@ -436,8 +436,30 @@ public partial class ShellViewModel : ObservableObject,
 
             case CommandResultKind.GoHome:
                 {
-                    // Go back to the main page, but keep it open
-                    GoHome();
+                    if (onBeforeShowConfirmation is not null)
+                    {
+                        // Dock context: show the palette window at the correct
+                        // position, then perform a full reset to the root page
+                        // (the navigation stack may be empty if the palette was
+                        // previously hidden).
+                        try
+                        {
+                            onBeforeShowConfirmation.Invoke();
+                        }
+                        catch (Exception ex)
+                        {
+                            CoreLogger.LogError(ex.ToString());
+                        }
+
+                        ResetToHome();
+                    }
+                    else
+                    {
+                        // Non-dock context: palette is already visible, just
+                        // pop the navigation stack back to root.
+                        GoHome();
+                    }
+
                     break;
                 }
