+++ src/modules/cmdpal/Tests/Microsoft.CmdPal.UI.ViewModels.UnitTests/ExtensionGalleryViewModelTests.cs
@@ -13,6 +13,7 @@ using Microsoft.CmdPal.Common.Services;
 using Microsoft.CmdPal.Common.WinGet.Models;
 using Microsoft.CmdPal.Common.WinGet.Services;
 using Microsoft.CmdPal.UI.ViewModels.Gallery;
+using Microsoft.CmdPal.UI.ViewModels.Services;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
@@ -61,7 +62,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory());
 
@@ -118,7 +119,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory(winGetService.Object, winGetStatusService.Object),
             winGetService.Object,
@@ -202,7 +203,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory(winGetService.Object, winGetStatusService.Object),
             winGetService.Object,
@@ -250,7 +251,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory());
 
@@ -292,7 +293,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory());
 
@@ -324,7 +325,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory());
 
@@ -359,7 +360,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory());
 
@@ -389,7 +390,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory());
 
@@ -421,7 +422,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory());
 
@@ -485,7 +486,7 @@ public class ExtensionGalleryViewModelTests
 
         using var viewModel = new ExtensionGalleryViewModel(
             galleryService.Object,
-            extensionService.Object,
+            new[] { extensionService.Object },
             NullLogger<ExtensionGalleryViewModel>.Instance,
             CreateGalleryExtensionViewModelFactory(winGetPackageStatusService: winGetStatusService.Object),
             winGetPackageManagerService: null,
